// frontend/src/app/swagger/model/instanceReadDto.ts
/**
 * S2Retro.Presentation
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InstancePhase } from './instancePhase';
import { InstanceRowReadDto } from './instanceRowReadDto';
import { InstanceColumnReadDto } from './instanceColumnReadDto';


export interface InstanceReadDto { 
    id?: string;
    name?: string | null;
    isArchived?: boolean;
    entryPhaseStart?: string;
    entryPhaseEnd?: string | null;
    phase?: InstancePhase;
    boardId?: string;
    columns?: Array<InstanceColumnReadDto> | null;
    rows?: Array<InstanceRowReadDto> | null;
}
export namespace InstanceReadDto {
}