// frontend/src/app/swagger/model/instanceColumnReadDto.ts
/**
 * S2Retro.Presentation
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ColumnType } from './columnType';
import { InstanceCategoryReadDto } from './instanceCategoryReadDto';


export interface InstanceColumnReadDto { 
    id?: string;
    position?: number;
    name?: string | null;
    type?: ColumnType;
    instanceId?: string;
    instanceCategoryId?: string | null;
    category?: InstanceCategoryReadDto;
}
export namespace InstanceColumnReadDto {
}